fix: chapterID -> autobiographyID로 n번째 자서전 이동하는 로직 수정

+++ lib/models/home/autobiography.dart
@@ -10,7 +10,7 @@ class HomeAutobiography {
 
   factory HomeAutobiography.fromJson(Map<String, dynamic> json) {
     return HomeAutobiography(
-      autobiographyId: json['autobiographyId'],
+      autobiographyId: json['autobiographyId'] as int,
       title: json['title'],
       contentPreview: json['contentPreview'],
       coverImageUrl: json['coverImageUrl'],
